OPMPHM: rename opmphm to kdbopmphm
